猫が並んだ。

+++ Sources/RunCat/Point.swift
@@ -1,6 +1,6 @@
 struct Point {
-    var x: Int
-    var y: Int
+    var x: UInt16
+    var y: UInt16
 
     static let zero = Point(x: .zero, y: .zero)
 }
