+++ Sources/RunCat/Size.swift
@@ -1,6 +1,6 @@
 struct Size {
-    var width: Int
-    var height: Int
+    var width: UInt16
+    var height: UInt16
 
     static let screen = Size(width: 240, height: 160)
 }
